Always \0 terminate string type attributes in mschap_add_reply()

+++ src/modules/rlm_mschap/mschap.h
@@ -16,7 +16,7 @@ void mschap_auth_response(char const *username,
 			  uint8_t const *peer_challenge, uint8_t const *auth_challenge,
 			  char *response);
 void mschap_add_reply(REQUEST *request, unsigned char ident,
-		      char const *name, char const *value, int len);
+		      char const *name, char const *value, size_t len);
 
 
 #endif /*_MSCHAP_H*/
